+++ template/render_test.go
@@ -8,8 +8,10 @@ import (
 )
 
 const (
-	testJobName = "levantExample"
-	testDCName  = "dc13"
+	testJobName           = "levantExample"
+	testJobNameOverwrite  = "levantExampleOverwrite"
+	testJobNameOverwrite2 = "levantExampleOverwrite2"
+	testDCName            = "dc13"
 )
 
 func TestTemplater_RenderTemplate(t *testing.T) {
@@ -21,7 +23,7 @@ func TestTemplater_RenderTemplate(t *testing.T) {
 	fVars := make(map[string]string)
 
 	// Test basic TF template render.
-	job, err = RenderJob("test-fixtures/single_templated.nomad", "test-fixtures/test.tf", "", &fVars)
+	job, err = RenderJob("test-fixtures/single_templated.nomad", []string{"test-fixtures/test.tf"}, "", &fVars)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -30,7 +32,7 @@ func TestTemplater_RenderTemplate(t *testing.T) {
 	}
 
 	// Test basic YAML template render.
-	job, err = RenderJob("test-fixtures/single_templated.nomad", "test-fixtures/test.yaml", "", &fVars)
+	job, err = RenderJob("test-fixtures/single_templated.nomad", []string{"test-fixtures/test.yaml"}, "", &fVars)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -38,8 +40,36 @@ func TestTemplater_RenderTemplate(t *testing.T) {
 		t.Fatalf("expected %s but got %v", testJobName, *job.Name)
 	}
 
+	// Test multiple var-files
+	job, err = RenderJob("test-fixtures/single_templated.nomad", []string{"test-fixtures/test.yaml", "test-fixtures/test-overwrite.yaml"}, "", &fVars)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if *job.Name != testJobNameOverwrite {
+		t.Fatalf("expected %s but got %v", testJobNameOverwrite, *job.Name)
+	}
+
+	// Test multiple var-files of different types
+	job, err = RenderJob("test-fixtures/single_templated.nomad", []string{"test-fixtures/test.tf", "test-fixtures/test-overwrite.yaml"}, "", &fVars)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if *job.Name != testJobNameOverwrite {
+		t.Fatalf("expected %s but got %v", testJobNameOverwrite, *job.Name)
+	}
+
+	// Test multiple var-files with var-args
+	fVars["job_name"] = testJobNameOverwrite2
+	job, err = RenderJob("test-fixtures/single_templated.nomad", []string{"test-fixtures/test.tf", "test-fixtures/test-overwrite.yaml"}, "", &fVars)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if *job.Name != testJobNameOverwrite2 {
+		t.Fatalf("expected %s but got %v", testJobNameOverwrite2, *job.Name)
+	}
+
 	// Test empty var-args and empty variable file render.
-	job, err = RenderJob("test-fixtures/none_templated.nomad", "", "", &fVars)
+	job, err = RenderJob("test-fixtures/none_templated.nomad", []string{}, "", &fVars)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -48,8 +78,9 @@ func TestTemplater_RenderTemplate(t *testing.T) {
 	}
 
 	// Test var-args only render.
+	delete(fVars, "job_name")
 	fVars["job_name"] = testJobName
-	job, err = RenderJob("test-fixtures/single_templated.nomad", "", "", &fVars)
+	job, err = RenderJob("test-fixtures/single_templated.nomad", []string{}, "", &fVars)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -60,7 +91,7 @@ func TestTemplater_RenderTemplate(t *testing.T) {
 	// Test var-args and variables file render.
 	delete(fVars, "job_name")
 	fVars["datacentre"] = testDCName
-	job, err = RenderJob("test-fixtures/multi_templated.nomad", "test-fixtures/test.yaml", "", &fVars)
+	job, err = RenderJob("test-fixtures/multi_templated.nomad", []string{"test-fixtures/test.yaml"}, "", &fVars)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -73,7 +104,7 @@ func TestTemplater_RenderTemplate(t *testing.T) {
 
 	// Test var-args only render.
 	fVars["job_name"] = testJobName
-	_, err = RenderTemplate("test-fixtures/missing_var.nomad", "", "", &fVars)
+	_, err = RenderTemplate("test-fixtures/missing_var.nomad", []string{}, "", &fVars)
 	if err == nil {
 		t.Fatal("expected err to not be nil")
 	}